Retain the first link encountered in target text.
Given slabText is intended for headings, it's relatively unlikely there'd be multiple links, but just in case.
I wasn't sure whether this should be a setting, and it didn't matter in my scenario, so I just left that out. Should be trivial to make it one, though.
You can see this in action at betteridgeslaw.com